requirements.txt
# Core
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# API & Modal
fastapi>=0.104.0
uvicorn>=0.24.0  # Para correr FastAPI localmente
modal>=0.63.0
requests>=2.31.0

# LLM
openai>=1.0.0             # Para OpenAI API (recomendado)
llama-cpp-python>=0.2.90  # Para LLM local (Modal y Local)
huggingface-hub>=0.20.0   # Para descargar modelos
ollama>=0.1.0             # Para text-to-SQL local (opcional)

# Telegram
python-telegram-bot>=20.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
